# server/src/main/resources/swagger.yaml
openapi: 3.0.0
info:
  title: Documentación API de Productos
  version: 1.0.0
  description: API para gestionar productos
paths:
...
  /facturas:
    get:
      summary: Obtener todas las facturas
      responses:
        '200':
          description: Lista de facturas obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Factura'
    post:
      summary: Crear una nueva factura
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Factura'
      responses:
        '200':
          description: Factura creada exitosamente
          content:
            text/plain:
              schema:
                type: string
  /facturas/{id}:
    get:
      summary: Obtener una factura por ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la factura a obtener
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Factura obtenida exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Factura'
    delete:
      summary: Eliminar una factura por ID
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la factura a eliminar
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Factura eliminada exitosamente
          content:
            text/plain:
              schema:
                type: string
...
  /facturas/{id}/producto/{productId}:
    post:
      summary: Agregar producto a una factura
      parameters:
        - name: id
          in: path
          required: true
          description: ID de la factura a la que se agregará el producto
          schema:
            type: integer
            format: int64
        - name: productId
          in: path
          required: true
          description: ID del producto a agregar
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Producto agregado exitosamente a la factura
          content:
            text/plain:
              schema:
                type: string
  /facturas/cliente/{dni}:
    get:
      summary: Obtener facturas de un cliente por DNI
      parameters:
        - name: dni
          in: path
          required: true
          description: DNI del cliente
          schema:
            type: string
      responses:
        '200':
          description: Facturas obtenidas exitosamente para el cliente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Factura'
...
components:
  schemas:
...
    Factura:
      type: object
      properties:
        id:
          type: integer
          format: int64
...